[PRISM] Move scope_node itself to CRuby, create prism_compile.h
