+++ iseq.c
@@ -43,7 +43,7 @@
 #include "builtin.h"
 #include "insns.inc"
 #include "insns_info.inc"
-#include "prism/prism.h"
+#include "prism_compile.h"
 
 VALUE rb_cISeq;
 static VALUE iseqw_new(const rb_iseq_t *iseq);
@@ -1390,8 +1390,6 @@ iseqw_s_compile(int argc, VALUE *argv, VALUE self)
 
     return iseqw_new(rb_iseq_compile_with_option(src, file, path, line, opt));
 }
-
-void pm_scope_node_init(const pm_node_t *node, pm_scope_node_t *scope, pm_scope_node_t *previous, pm_parser_t *parser);
 
 static VALUE
 iseqw_s_compile_prism(int argc, VALUE *argv, VALUE self)
